+++ package.json
@@ -1,37 +1,43 @@
 {
-  "name": "my-preact-app",
-  "version": "1.0.0",
-  "description": "A Preact app",
-  "main": "index.js",
-  "scripts": {
-    "start": "webpack serve --mode development --open --hot",
-    "build": "webpack --mode production",
-    "test": "jest --silent=false"
-  },
+  "name": "pdf-flipper",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.15.6",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "pdf-lib": "^1.17.1",
-    "preact": "^10.5.14"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-pdf": "^7.7.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
-  "devDependencies": {
-    "@babel/core": "^7.13.15",
-    "@babel/plugin-transform-modules-commonjs": "^7.23.3",
-    "@babel/plugin-transform-react-jsx": "^7.23.4",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-react": "^7.23.3",
-    "@testing-library/jest-dom": "^6.2.0",
-    "@testing-library/preact": "^3.2.3",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^6.9.1",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.1",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.0.0",
-    "jest-environment-jsdom": "^26.0.0",
-    "less": "^4.2.0",
-    "less-loader": "^12.0.0",
-    "style-loader": "^3.3.4",
-    "webpack": "^5.36.2",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
