--- pdf-flipper/src/components/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import Main from './Main';
-
-export default class App extends React.Component {
-    render() {
-        return (
-            <div>
-                <Main />
-            </div>
-        );
-    }
-}
