--- pdf-flipper/test/components/app.test.js
@@ -1,11 +1,0 @@
-import { h } from 'preact';
-import { render } from '@testing-library/preact';
-
-import App from '../../src/components/App';
-
-describe('App component', () => {
-  test('renders without crashing', () => {
-    const { container } = render(<App />);
-    expect(container).toBeTruthy();
-  });
-});
