--- pdf-flipper/test/jest.setup.js
@@ -1,2 +1,0 @@
-// setup
-
