--- src/components/styles.less
@@ -1,9 +1,0 @@
-
-.appMainBody {
-    color: #DEDEDE;
-    font-family: "Open Sans", Helvetica, sans-serif;
-}
-
-.main {
-    border-radius: 10px;
-}
