--- webpack.config.js
@@ -1,70 +1,0 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
-module.exports = {
-  entry: './src/index.js',
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    filename: 'bundle.js'
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader'
-        }
-      },
-      {
-        test: /\.less$/,
-        use: [
-            'style-loader',
-            {
-                loader: 'css-loader',
-                options: {
-                    modules: {
-                        localIdentName: '[name]__[local]___[hash:base64:5]'
-                    },
-                }
-            },
-            'less-loader'
-        ]
-    },
-    {
-        test: /\.css$/,
-        use: [
-            'style-loader',
-            'css-loader'
-        ]
-    },
-    {
-        test: /\.(png|jpe?g|gif|svg)$/i,
-        use: [
-            {
-                loader: 'file-loader',
-                options: {
-                    name: '[path][name].[ext]',
-                },
-            },
-        ],
-    }
-    ]
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './public/index.html'
-    })
-  ],
-  devServer: {
-    contentBase: './dist',
-    hot: true
-  },
-  resolve: { // added this resolve section for the preact/compat alias
-    alias : {
-        "react": "preact/compat",
-        "react-dom/test-utils": "preact/test-utils",
-        "react-dom": "preact/compat",
-    }
-  }
-};
